@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
-import rootReducer from '../rootReducer';
-import rootSaga from '../rootSaga';
+import rootReducer from 'reducers/root';
+import rootSaga from 'sagas/root';
 
 const sagaMiddleware = createSagaMiddleware(rootSaga);
 
@@ -9,7 +9,7 @@ export default function configureStore(initialState) {
   const store = createStore(
     rootReducer,
     initialState,
-    applyMiddleware(sagaMiddleware)
+    applyMiddleware(sagaMiddleware),
   );
 
   return store;
